analytics-engine/requirements.txt
kafka-python
neo4j
pandas